backend/requirements.txt
flask==3.1.2
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
faiss-cpu==1.12.0
sentence-transformers==5.1.2
transformers==4.57.1
torch==2.9.0
pyyaml==6.0.3
requests==2.32.5
